# api/docs/admin/admin_swagger.yaml
basePath: /api
definitions:
...
  model.AddressBookCollection:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    required:
    - name
    type: object
  model.AddressBookCollectionList:
    properties:
      list:
        items:
          $ref: '#/definitions/model.AddressBookCollection'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  model.AddressBookCollectionRule:
    properties:
      collection_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      rule:
        description: '0: 无 1: 读 2: 读写  3: 完全控制'
        maximum: 3
        minimum: 1
        type: integer
      to_id:
        type: integer
      type:
        description: '1: 个人 2: 群组'
        maximum: 2
        minimum: 1
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    required:
    - collection_id
    - rule
    - to_id
    - type
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: 接口
  title: 管理系统API
  version: "1.0"
paths:
...
  /admin/address_book_collection/create:
    post:
      consumes:
      - application/json
      description: 创建地址簿名称
      parameters:
      - description: 地址簿名称信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AddressBookCollection'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AddressBookCollection'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 创建地址簿名称
      tags:
      - 地址簿名称
  /admin/address_book_collection/delete:
    post:
      consumes:
      - application/json
      description: 地址簿名称删除
      parameters:
      - description: 地址簿名称信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AddressBookCollection'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 地址簿名称删除
      tags:
      - 地址簿名称
  /admin/address_book_collection/detail/{id}:
    get:
      consumes:
      - application/json
      description: 地址簿名称详情
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AddressBookCollection'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 地址簿名称详情
      tags:
      - 地址簿名称
  /admin/address_book_collection/list:
    get:
      consumes:
      - application/json
      description: 地址簿名称列表
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 页大小
        in: query
        name: page_size
        type: integer
      - description: 是否是我的
        in: query
        name: is_my
        type: integer
      - description: 用户id
        in: query
        name: user_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AddressBookCollectionList'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 地址簿名称列表
      tags:
      - 地址簿名称
  /admin/address_book_collection/update:
    post:
      consumes:
      - application/json
      description: 地址簿名称编辑
      parameters:
      - description: 地址簿名称信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AddressBookCollection'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AddressBookCollection'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 地址簿名称编辑
      tags:
      - 地址簿名称
  /admin/address_book_collection_rule/create:
    post:
      consumes:
      - application/json
      description: 创建地址簿规则
      parameters:
      - description: 地址簿规则信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AddressBookCollectionRule'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AddressBookCollection'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 创建地址簿规则
      tags:
      - 地址簿规则
  /admin/address_book_collection_rule/delete:
    post:
      consumes:
      - application/json
      description: 地址簿规则删除
      parameters:
      - description: 地址簿规则信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AddressBookCollectionRule'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 地址簿规则删除
      tags:
      - 地址簿规则
  /admin/address_book_collection_rule/detail/{id}:
    get:
      consumes:
      - application/json
      description: 地址簿规则详情
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AddressBookCollectionRule'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 地址簿规则详情
      tags:
      - 地址簿规则
  /admin/address_book_collection_rule/list:
    get:
      consumes:
      - application/json
      description: 地址簿规则列表
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 页大小
        in: query
        name: page_size
        type: integer
      - description: 是否是我的
        in: query
        name: is_my
        type: integer
      - description: 用户id
        in: query
        name: user_id
        type: integer
      - description: 地址簿集合id
        in: query
        name: collection_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AddressBookCollectionList'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 地址簿规则列表
      tags:
      - 地址簿规则
  /admin/address_book_collection_rule/update:
    post:
      consumes:
      - application/json
      description: 地址簿规则编辑
      parameters:
      - description: 地址簿规则信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AddressBookCollectionRule'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AddressBookCollection'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 地址簿规则编辑
      tags:
      - 地址簿规则
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  token:
    in: header
    name: api-token
    type: apiKey
swagger: "2.0"